# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Internity API
  version: 1.0.0
  description: |
    Internity is a school internship (PKL — Praktik Kerja Lapangan) management system.
    It coordinates the full lifecycle of a student's internship: browsing and applying to
    company-posted vacancies, the accept/reject workflow, day-to-day attendance and journal
    tracking once placed, mentor/coordinator scoring, certificate generation, coordinator
    site-visit monitoring, company/mentor reviews, and school/department/company/course
    org-structure administration.

    ## Response envelope
    Every JSON response (except binary downloads, see below) uses one consistent envelope.

    Success:
    ```json
    { "success": true, "data": { ... }, "message": "OK", "meta": { "request_id": "...", "pagination": null } }
    ```

    Error:
    ```json
    { "success": false, "data": null, "message": "...", "error": { "code": "VALIDATION_ERROR", "details": [] }, "meta": { "request_id": "..." } }
    ```

    `error.code` is one of `BAD_REQUEST` (400), `VALIDATION_ERROR` (422), `UNAUTHENTICATED` (401),
    `FORBIDDEN` (403), `NOT_FOUND` (404), `CONFLICT` (409), `RATE_LIMITED` (429), `INTERNAL_ERROR` (500).

    A small number of endpoints intentionally return **no body at all** — `204 No Content`
    (every `DELETE`, plus unsave-vacancy) — these are called out explicitly rather than
    documented as an empty envelope. Certificate and export endpoints stream raw
...

    ## Authentication
...

...
servers:
...
security:
  - cookieAuth: []
paths:
...
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Rotate the session using the refresh cookie
      description: >
        Public endpoint (reads the `internity_refresh` cookie itself, not the CSRF/session
        cookie). Reuse of an already-rotated refresh token is treated as theft evidence and
        revokes every session in that token family, forcing a full re-login.
      security: []
      responses:
        '200':
          description: Session refreshed
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties: { data: { $ref: '#/components/schemas/UserWrapper' } }
        '401':
          description: Missing, invalid, or reused/expired refresh token
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
              example: { success: false, data: null, message: "Session expired, please log in again", error: { code: "UNAUTHENTICATED" }, meta: {} }
        '403':
          description: Account deactivated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: internity_session
      description: >
        HttpOnly session cookie set by /auth/login, /auth/register, or /auth/refresh.
        Mutating requests additionally require the X-CSRF-Token header — see the CSRFToken parameter.
...
  schemas:
...
    Pagination:
      type: object
      required: [page, limit, total]
      properties:
        page: { type: integer }
        limit: { type: integer }
        total: { type: integer, format: int64 }
    Meta:
      type: object
      properties:
        request_id: { type: string }
        pagination:
          anyOf:
            - $ref: '#/components/schemas/Pagination'
            - type: "null"
    SuccessEnvelope:
      type: object
      required: [success, data, message]
      description: The success shape every non-binary endpoint returns, with `data` typed per-endpoint via `allOf` where documented.
      properties:
        success: { type: boolean, const: true }
        data: {}
        message: { type: string }
        meta: { $ref: '#/components/schemas/Meta' }
    ErrorCode:
      type: string
      enum: [BAD_REQUEST, VALIDATION_ERROR, UNAUTHENTICATED, FORBIDDEN, NOT_FOUND, CONFLICT, RATE_LIMITED, INTERNAL_ERROR]
    ErrorDetail:
      type: object
      required: [issue]
      properties:
        field: { type: string }
        issue: { type: string }
    ErrorEnvelope:
      type: object
      required: [success, data, message, error]
      properties:
        success: { type: boolean, const: false }
        data: { type: "null" }
        message: { type: string }
        error:
          type: object
          required: [code]
          properties:
            code: { $ref: '#/components/schemas/ErrorCode' }
            details: { type: array, items: { $ref: '#/components/schemas/ErrorDetail' } }
        meta: { $ref: '#/components/schemas/Meta' }

    # ---------- Identity ----------
    UserResponse:
      type: object
      required: [id, role, name, email, is_active, created_at]
      properties:
        id: { type: string, format: uuid }
        role: { type: string, enum: [admin, coordinator, mentor, student] }
        school_id: { type: [integer, "null"], format: int64 }
        department_id: { type: [integer, "null"], format: int64 }
        company_id: { type: [integer, "null"], format: int64 }
        course_id: { type: [integer, "null"], format: int64 }
        name: { type: string }
        email: { type: string, format: email }
        nis: { type: [string, "null"], description: Student ID number. }
        gender: { type: [string, "null"], enum: [male, female, null] }
        bio: { type: [string, "null"] }
        address: { type: [string, "null"] }
        phone: { type: [string, "null"] }
        date_of_birth: { type: [string, "null"], format: date-time }
        avatar_key: { type: [string, "null"], description: Object storage key for the avatar image. }
        resume_key: { type: [string, "null"], description: Object storage key for the uploaded resume. }
        skills: { type: [string, "null"] }
        is_active: { type: boolean }
        created_at: { type: string, format: date-time }
    UserWrapper:
      type: object
      required: [user]
      properties:
        user: { $ref: '#/components/schemas/UserResponse' }
...